iso27001-fastapi/pyproject.toml
[project]
name = "iso27001-fastapi"
version = "1.0.0"
description = "ISO 27001 Compliant REST API"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.25",
    "pydantic[email]>=2.6.0",
    "pydantic-settings>=2.1.0",
    "PyJWT>=2.8.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=3.2,<4.0",
    "python-multipart>=0.0.9",
    "prometheus-client>=0.19.0",
    "redis>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
    "mypy>=1.8.0",
    "types-passlib>=1.7.0",
    "import-linter>=2.1",
]
aws = [
    "boto3>=1.34",
    "aws-xray-sdk>=2.12",
]

[tool.mypy]
python_version = "3.11"
strict = true
exclude = ["tests/"]

[[tool.mypy.overrides]]
module = "app.infrastructure.aws_telemetry"
warn_unused_ignores = false